requirements.txt
asgiref==3.7.2
Django==4.2.4
django-environ==0.10.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
phonenumberslite==8.13.18
psycopg2-binary==2.9.7
pytz==2023.3
sqlparse==0.4.4
setuptools==65.5.1
typing_extensions==4.7.1
